Company

+++ src/models/bm.documents.model.js
@@ -958,12 +958,12 @@ export async function createDocumentFromProject(
           pm.quantity,
           CASE
             WHEN p.default_pricing = true THEN COALESCE(pm.sell_cost_override, pm.unit_cost_override, 0)
-            ELSE COALESCE(ROUND((pm.unit_cost_override * (1 + $4))::numeric, 2), 0)
+            ELSE COALESCE(ROUND((pm.unit_cost_override * (1::numeric + $4::numeric))::numeric, 2), 0)
           END AS unit_price,
           ROUND(
             (pm.quantity * CASE
               WHEN p.default_pricing = true THEN COALESCE(pm.sell_cost_override, pm.unit_cost_override, 0)
-              ELSE COALESCE((pm.unit_cost_override * (1 + $4)), 0)
+              ELSE COALESCE((pm.unit_cost_override * (1::numeric + $4::numeric)), 0)
             END)::numeric,
             2
           ) AS line_total
@@ -996,12 +996,12 @@ export async function createDocumentFromProject(
           l.unit_type,
           CASE
             WHEN p.default_pricing = true THEN COALESCE(pl.sell_cost_override, l.sell_cost, l.unit_cost, 0)
-            ELSE COALESCE(ROUND((COALESCE(pl.unit_cost_override, l.unit_cost) * (1 + $4))::numeric, 2), 0)
+            ELSE COALESCE(ROUND((COALESCE(pl.unit_cost_override, l.unit_cost) * (1::numeric + $4::numeric))::numeric, 2), 0)
           END AS unit_price,
           ROUND(
             (pl.quantity * CASE
               WHEN p.default_pricing = true THEN COALESCE(pl.sell_cost_override, l.sell_cost, l.unit_cost, 0)
-              ELSE COALESCE((COALESCE(pl.unit_cost_override, l.unit_cost) * (1 + $4)), 0)
+              ELSE COALESCE((COALESCE(pl.unit_cost_override, l.unit_cost) * (1::numeric + $4::numeric)), 0)
             END)::numeric,
             2
           ) AS line_total
